fix: back rebound effect( in Swipe functionality ) not showing the content of itemView appropriately
After invoking notifyItemRangeChanged(in onItemDismiss), the next time there is a swipe, the back rebound effect(don't know where is done, suspect that in onDrawChild because there a calls to this function with dx=1080) force to show incorrect views in RecyclerView item. In order for the views not to shows, change explicitly the visibility of them. The rebound effect normally makes this views invisible, but not if there has been a  notifyItemRangeChanged before.