fix: Statusbar shadow at the bottom, very annoying effect.
This happens when altering the elevation of the AppBarLayout and the statusbarColor is set to transparent in styles. Sometimes we can simply delete this line in styles. Others we cannot.
One approach is to change from true to false the following property in CoordnatorLayout: fitsSystemWindows=false. This way we can change the AppBarLayout elevation without impacting on the status bar.