feat:  implement scrolling behavior for AppBarLayout and afterwards quick return pattern with bottom navigation view.
The initial problem was that in activity_main.xml,  AppBarLayout was "wrapped" by a RelativeLayout. This prevented the scrolling flag set in AppBarLayout to work. One solution is to add that flag also to RelativeLayout or add a CollapsingToolbarLayout. Either alternative was practically the same, since RelativeLayout worked as a FrameLayout because there wasn't aligment or positioning based on the other views. I opted for CollapsingToolbarLayout since it is the standart Android way. Both alternatives provoked a change in the layout of fragment_tab_map(only in this layout) and the bottom margin wasn't taked into account, no matter the value set, causing an overlap of the bottom sheet with the bottom navigation. To solve this issue wrap the bottom sheet inside another CoordinatorLayout and set the margin there.  The implementation of quick return pattern was relatively straightforward, just take into account that now there is no longer a need for a bottom margin in the layouts with a recyclerview.