fix: snackbar now slides up to the correct position and does it behind the bottom navigation.
There was a in incorrect property in activity_main layout. Bottom navigation was using  layout_gravity(from LinearLayout) instead of a relative param to position itself at the bottom of the screen. That way the "snack coordinator" was wong positioned, behind the bottom navigation istead of above.

Change accent color to blue.