feat: enable animations when loading the adapter in ListTabFragment.
If a new location arrives, the user will be presented with a snackbar to decide whether or not he wants to update
the data. If the answer is yes, the adapter is then filled with new data. This commit enable that in this process the
animations are enabled. Since now the ui thread doesn't have so much cpu pressure, it is ok to do that.